+++ Architectures/Project.toml
@@ -1,0 +1,10 @@
+name = "Architectures"
+uuid = "c5c4acf7-95e6-4f60-ae13-83e51f04c294"
+authors = ["SpeedyWeather contributors"]
+version = "0.1.0"
+
+[deps]
+KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
+
+[compat]
+KernelAbstractions = "0.9.38"
