+++ LowerTriangularArrays/src/LowerTriangularArrays.jl
@@ -8,7 +8,7 @@ import Adapt: Adapt, adapt
 import GPUArrays
 import KernelAbstractions
 
-include("../Architectures.jl")
+include("../../Architectures/src/Architectures.jl")
 import .Architectures
 import .Architectures: AbstractArchitecture, on_architecture, 
     array_type, ismatching, CPU, GPU, architecture, nonparametric_type
