+++ RingGrids/src/utility_functions.jl
