+++ SpeedyTransforms/src/aliasing.jl
