+++ SpeedyTransforms/src/legendre_ka.jl
