+++ SpeedyTransforms/src/legendre_shortcuts.jl
