+++ SpeedyTransforms/src/power_spectrum.jl
