+++ SpeedyTransforms/src/scratch_memory.jl
