+++ SpeedyTransforms/src/show.jl
