+++ SpeedyTransforms/src/spectral_gradients.jl
