+++ SpeedyTransforms/src/spectral_transform.jl
