+++ SpeedyTransforms/src/spectral_truncation.jl
