+++ Utils/Project.toml
@@ -1,0 +1,22 @@
+name = "Utils"
+uuid = "c092a71f-dc30-4948-abbf-534dc63e7c92"
+authors = ["SpeedyWeather contributors"]
+version = "0.1.0"
+
+[deps]
+ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
+ConstructionBase = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
+DomainSets = "5b8099bc-c8ec-5219-889f-1d9e522a28bf"
+MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
+ModelParameters = "4744a3fa-6c31-4707-899e-a3298e4618ad"
+
+[compat]
+ComponentArrays = "0.15.29"
+ConstructionBase = "1.6.0"
+Dates = "1.11.0"
+DocStringExtensions = "0.9.5"
+DomainSets = "0.7.16"
+MacroTools = "0.5.16"
+ModelParameters = "0.4.5"
